Fixed some FindBugs warnings in the cache tests

+++ hazelcast/src/test/java/com/hazelcast/cache/CacheReadWriteThroughTest.java
@@ -342,34 +342,29 @@ public class CacheReadWriteThroughTest extends HazelcastTestSupport {
         }
 
         @Override
-        public void write(Cache.Entry<? extends Integer, ? extends Integer> entry)
-                throws CacheWriterException {
-            Integer value = entry.getValue().intValue();
+        public void write(Cache.Entry<? extends Integer, ? extends Integer> entry) throws CacheWriterException {
+            Integer value = entry.getValue();
             if (!isAcceptableValue(value)) {
                 throw new CacheWriterException("Value is invalid: " + value);
             }
         }
 
         @Override
-        public void writeAll(Collection<Cache.Entry<? extends Integer, ? extends Integer>> entries)
-                throws CacheWriterException {
+        public void writeAll(Collection<Cache.Entry<? extends Integer, ? extends Integer>> entries) throws CacheWriterException {
         }
 
         @Override
-        public void delete(Object key)
-                throws CacheWriterException {
+        public void delete(Object key) throws CacheWriterException {
         }
 
         @Override
-        public void deleteAll(Collection<?> keys)
-                throws CacheWriterException {
+        public void deleteAll(Collection<?> keys) throws CacheWriterException {
         }
     }
 
     public interface ValueChecker extends Serializable {
 
         boolean isAcceptableValue(int value);
-
     }
 
     public static class ModValueChecker implements ValueChecker {
@@ -384,7 +379,6 @@ public class CacheReadWriteThroughTest extends HazelcastTestSupport {
         public boolean isAcceptableValue(int value) {
             return value % mod != 0;
         }
-
     }
 
     public static class MaxValueChecker implements ValueChecker {
